refactor: 주문 테이블(Order Table) 관련 클래스를 주문 테이블 패키지 아래에 추가.
