+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.domain;
 
 import kitchenpos.order.domain.enums.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.orderTable.domain.OrderTable;
 import kitchenpos.order.domain.enums.OrderType;
 
 import javax.persistence.*;
