+++ src/main/java/kitchenpos/order/service/OrderService.java
@@ -1,6 +1,5 @@
 package kitchenpos.order.service;
 
-import kitchenpos.domain.*;
 import kitchenpos.infra.KitchenridersClient;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.repository.MenuRepository;
@@ -9,6 +8,8 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.enums.OrderStatus;
 import kitchenpos.order.domain.enums.OrderType;
 import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.orderTable.domain.OrderTable;
+import kitchenpos.orderTable.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
