+++ src/main/java/kitchenpos/orderTable/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.orderTable.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
