+++ src/main/java/kitchenpos/orderTable/repository/JpaOrderTableRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.orderTable.repository;
 
+import kitchenpos.orderTable.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
